fixed dangerous wind service tests
